Deduplication fixes
